Add support for service accounts in fissile (for use with RBAC)
This adds support for kube RBAC authorization to have custom role
bindings for the default service accounts in the (fissile) roles.  This
is needed as otherwise configgin will be unable to enumerate other pods
in the namespace and to patch the context pod to export BOSH properties.

This also turns on RBAC for the vagrant box so the new code gets tested
properly. (The generated charts should also work with non-RBAC clusters
as they're all guarded by configuration values in values.yaml.)